refactor: consolidate useAdminPersonas into useAgents as useAdminAgents
Move useAdminPersonas into useAgents.ts and rename to useAdminAgents.
Delete the standalone useAdminPersonas.ts file. Update all consumers.
Also move PAGE_SIZE to lib/constants.ts as DEFAULT_PAGE_SIZE.
