+++ web/src/app/admin/discord-bot/[guild-id]/page.tsx
@@ -24,7 +24,7 @@ import {
 } from "@/app/admin/discord-bot/lib";
 import { DiscordChannelsTable } from "@/app/admin/discord-bot/[guild-id]/DiscordChannelsTable";
 import { DiscordChannelConfig } from "@/app/admin/discord-bot/types";
-import { useAdminPersonas } from "@/hooks/useAdminPersonas";
+import { useAdminAgents } from "@/hooks/useAgents";
 import { Persona } from "@/app/admin/agents/interfaces";
 
 interface Props {
@@ -157,7 +157,7 @@ export default function Page({ params }: Props) {
     error: channelsError,
     refreshChannels,
   } = useDiscordChannels(guildId);
-  const { personas, isLoading: personasLoading } = useAdminPersonas({
+  const { personas, isLoading: personasLoading } = useAdminAgents({
     includeDefault: true,
   });
   const [isUpdating, setIsUpdating] = useState(false);
