+++ web/src/lib/constants.ts
@@ -137,3 +137,5 @@ export const DEFAULT_CONTEXT_TOKENS = 120_000;
 export const MAX_CHUNKS_FED_TO_CHAT = 25;
 
 export const APP_SLOGAN = "Open Source AI Platform";
+
+export const DEFAULT_PAGE_SIZE = 10;
