+++ web/src/refresh-pages/admin/AgentsPage/AgentsTable.tsx
@@ -10,7 +10,7 @@ import InputTypeIn from "@/refresh-components/inputs/InputTypeIn";
 import type { MinimalUserSnapshot } from "@/lib/types";
 import AgentAvatar from "@/refresh-components/avatars/AgentAvatar";
 import type { MinimalPersonaSnapshot } from "@/app/admin/agents/interfaces";
-import { useAdminPersonas } from "@/hooks/useAdminPersonas";
+import { useAdminAgents } from "@/hooks/useAgents";
 import { toast } from "@/hooks/useToast";
 import AgentRowActions from "@/refresh-pages/admin/AgentsPage/AgentRowActions";
 import { updateAgentDisplayPriorities } from "@/refresh-pages/admin/AgentsPage/svc";
@@ -20,6 +20,7 @@ import { SvgActions, SvgUser } from "@opal/icons";
 import Popover, { PopoverMenu } from "@/refresh-components/Popover";
 import { LineItemButton } from "@opal/components";
 import { useUser } from "@/providers/UserProvider";
+import { DEFAULT_PAGE_SIZE } from "@/lib/constants";
 import useFilter from "@/hooks/useFilter";
 import useMcpServers from "@/hooks/useMcpServers";
 import {
@@ -163,8 +164,6 @@ function buildColumns(onMutate: () => void) {
 // Component
 // ---------------------------------------------------------------------------
 
-const PAGE_SIZE = 10;
-
 export default function AgentsTable() {
   const [searchTerm, setSearchTerm] = useState("");
   const { user } = useUser();
@@ -179,7 +178,7 @@ export default function AgentsTable() {
     new Set()
   );
 
-  const { personas, isLoading, error, refresh } = useAdminPersonas();
+  const { personas, isLoading, refresh } = useAdminAgents();
   const { mcpData } = useMcpServers();
 
   const columns = useMemo(() => buildColumns(refresh), [refresh]);
@@ -369,10 +368,10 @@ export default function AgentsTable() {
   // Reorder handler
   // ---------------------------------------------------------------------------
 
-  const handleReorder = async (
+  async function handleReorder(
     _orderedIds: string[],
     changedOrders: Record<string, number>
-  ) => {
+  ) {
     try {
       await updateAgentDisplayPriorities(changedOrders);
       refresh();
@@ -382,22 +381,13 @@ export default function AgentsTable() {
       );
       refresh();
     }
-  };
+  }
 
   if (isLoading) {
     return (
       <div className="flex justify-center py-12">
         <SimpleLoader className="h-6 w-6" />
       </div>
-    );
-  }
-
-  if (error) {
-    console.error("Failed to load agents:", error);
-    return (
-      <Text as="p" secondaryBody text03>
-        Failed to load agents. Please try refreshing the page.
-      </Text>
     );
   }
 
@@ -537,7 +527,7 @@ export default function AgentsTable() {
         data={agentRows}
         columns={columns}
         getRowId={(row) => String(row.id)}
-        pageSize={PAGE_SIZE}
+        pageSize={DEFAULT_PAGE_SIZE}
         searchTerm={searchTerm}
         draggable={{
           onReorder: handleReorder,
